Create code
In this example, we have a Reminder class that takes in a message and a due date (in milliseconds) when it is created. The setReminder method uses the setTimeout function to log the message to the console after the specified amount of time has passed. Two instances of the class are then created and their setReminder methods are called.